Add initial work to have a distributed client
