+++ bmemcached/client/__init__.py
@@ -1,6 +1,9 @@
 from bmemcached.client.constants import SOCKET_TIMEOUT
 
 from .replicating import ReplicatingClient
+from .distributed import DistributedClient
+
+__all__ = ('Client', 'ReplicatingClient', 'DistributedClient', )
 
 
 # Keep compatibility with old versions
