+++ setup.py
@@ -15,6 +15,8 @@ setup(
         'Programming Language :: Python :: 3.4',
     ],
     install_requires=[
-        'six'
+        'six',
+        'uhashring',
+        'typing'
     ]
 )
